requirements.txt
# LangChain Examples - Dependencies
# Use: pip install -r requirements.txt

# Core
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
langchain-text-splitters>=0.3.0
langchain-community>=0.3.0

# Environment
python-dotenv>=1.0.0

# Resolve conflict with rendercv (requires packaging>=25.0)
packaging>=25.0

# RAG (03_rag_pipeline.py)
langchain-pinecone>=0.2.0
pinecone-client>=3.0.0
faiss-cpu>=1.7.0  # Fallback when Pinecone has no indexes

# Agents (04_agents.py) - LLMMathChain uses numexpr
numexpr>=2.8.0
